Initial code (v3 OpenAI)
I followed the tutorial and documentation to get to this point but the app does not work. After some debugging and digging though the internet I found that the documentation that I was using was for version 3 of the OpenAI Node package. I have found the migration guide at https://github.com/openai/openai-node/discussions/217 and will be using the guide tomorrow to finish up the work and get the code running